# swagger.yaml
definitions:
...
  CategorySubscription:
    type: object
    properties:
      id:
        type: string
        format: uuid
      name:
        type: string
      slug:
        type: string
      thread_count:
        type: integer
      description:
        type: string
...
  SubscriptionRequest:
    type: object
    required:
      - category_id
    properties:
      category_id:
        type: string
        format: uuid
      password:
        type: string
        description: Optional password when joining a private category
...
  ErrorResponse:
    description: ErrorResponse defines the standard error response structure
    properties:
      code:
        type: string
        x-go-name: Code
      details:
        items:
          $ref: "#/definitions/ValidationError"
        type: array
        x-go-name: Details
      error:
        type: string
        x-go-name: Error
      message:
        type: string
        x-go-name: Message
      success:
        type: boolean
        x-go-name: Success
      timestamp:
        type: string
        x-go-name: Timestamp
    type: object
    x-go-package: github.com/purnama354/sejiwa-api/internal/dto
...
info:
  title: Sejiwa API
  description: Mental health discussion platform API
  version: "1.0.0"
paths:
...
  /api/v1/users/me/subscriptions:
    get:
      summary: List category subscriptions
      description: Get all categories the authenticated user is subscribed to
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of subscribed categories
          schema:
            type: array
            items:
              $ref: "#/definitions/CategorySubscription"
        "401":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
    post:
      summary: Subscribe to a category
      description: Subscribe the authenticated user to a specific category
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/SubscriptionRequest"
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          description: Already subscribed to this category
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/errorResponse"
    delete:
      summary: Unsubscribe from a category
      description: Unsubscribe the authenticated user from a specific category
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/UnsubscribeRequest"
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
  /api/v1/users/me/saved-threads:
    get:
      summary: List saved threads
      description: Get paginated list of threads saved by the authenticated user
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          type: integer
          minimum: 1
          default: 1
          description: Page number for pagination
        - in: query
          name: limit
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Number of items per page
      responses:
        "200":
          description: Paginated list of saved threads
          schema:
            $ref: "#/definitions/SavedThreadsResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
    post:
      summary: Save a thread
      description: Save a specific thread for the authenticated user
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/SaveThreadRequest"
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          description: Thread already saved
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          $ref: "#/responses/errorResponse"
    delete:
      summary: Unsave a thread
      description: Remove a thread from saved threads for the authenticated user
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/UnsaveThreadRequest"
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
  /admin/auth/create-admin:
    post:
      description: Create an admin account (Admin only)
      operationId: createAdmin
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/CreateAdminRequest"
      responses:
        "201":
          $ref: "#/responses/moderatorProfileResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - admin
  /admin/auth/create-moderator:
    post:
      description: Create a moderator account (Admin only)
      operationId: createModerator
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/CreateModeratorRequest"
      responses:
        "201":
          $ref: "#/responses/moderatorProfileResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - admin
  /auth/login:
    post:
      description: Login and obtain access/refresh tokens
      operationId: authLogin
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        "200":
          $ref: "#/responses/authResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "423":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - auth
  /auth/register:
    post:
      description: Register a new user account
      operationId: authRegister
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/RegisterRequest"
      responses:
        "201":
          $ref: "#/responses/authResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - auth
  /categories:
    get:
      description: List categories
      operationId: listCategories
      responses:
        "200":
          $ref: "#/responses/categoryListResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - categories
    post:
      description: Create category (Admin only)
      operationId: createCategory
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/CreateCategoryRequest"
      responses:
        "201":
          $ref: "#/responses/categoryResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - categories
  /categories/{id}:
    delete:
      description: Delete category (Admin only)
      operationId: deleteCategory
      parameters:
        - description: Category ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - categories
    get:
      description: Get category by ID
      operationId: getCategoryByID
      parameters:
        - description: Category ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/categoryResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - categories
    put:
      description: Update category (Admin only)
      operationId: updateCategory
      parameters:
        - description: Category ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/UpdateCategoryRequest"
      responses:
        "200":
          $ref: "#/responses/categoryResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - categories
  /categories/{id}/threads:
    get:
      description: List threads in a specific category
      operationId: getThreadsByCategory
      parameters:
        - description: Category ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - description: Page number
          format: int64
          in: query
          name: page
          type: integer
          x-go-name: Page
        - description: Page size
          format: int64
          in: query
          name: pageSize
          type: integer
          x-go-name: PageSize
      responses:
        "200":
          $ref: "#/responses/threadListResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - threads
...
  /moderation/actions:
    get:
      description: List moderation actions (Moderator/Admin)
      operationId: getModerationActions
      parameters:
        - description: Page number
          format: int64
          in: query
          name: page
          type: integer
          x-go-name: Page
        - description: Page size
          format: int64
          in: query
          name: pageSize
          type: integer
          x-go-name: PageSize
        - description: Filter by moderator ID
          in: query
          name: moderator_id
          type: string
          x-go-name: ModeratorID
        - description: Filter by action type
          in: query
          name: action
          type: string
          x-go-name: Action
        - description: Filter from date (RFC3339)
          in: query
          name: from_date
          type: string
          x-go-name: FromDate
        - description: Filter to date (RFC3339)
          in: query
          name: to_date
          type: string
          x-go-name: ToDate
      responses:
        "200":
          $ref: "#/responses/moderationActionListResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/moderator-notes/{noteId}:
    delete:
      description: Delete a moderator note (Moderator/Admin)
      operationId: deleteModeratorNote
      parameters:
        - description: Note ID
          in: path
          name: noteId
          required: true
          type: string
          x-go-name: NoteID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/replies/{id}:
    post:
      description: Moderate a reply (Moderator/Admin)
      operationId: moderateReply
      parameters:
        - description: Reply ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/reports:
    get:
      description: List reports for moderation review (Moderator/Admin)
      operationId: getReportsForModeration
      parameters:
        - description: Page number (default 1)
          format: int64
          in: query
          name: page
          type: integer
          x-go-name: Page
        - description: Page size (default varies, use pageSize or limit)
          format: int64
          in: query
          name: pageSize
          type: integer
          x-go-name: PageSize
        - description: Limit (for some endpoints using 'limit' instead of 'pageSize')
          format: int64
          in: query
          name: limit
          type: integer
          x-go-name: Limit
        - description: Filter by status (pending, reviewed, resolved)
          in: query
          name: status
          type: string
          x-go-name: Status
        - description: Filter by priority (low, medium, high, critical)
          in: query
          name: priority
          type: string
          x-go-name: Priority
        - description: Optional category filter
          in: query
          name: category
          type: string
          x-go-name: Category
      responses:
        "200":
          $ref: "#/responses/reportListResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/reports/{id}:
    get:
      description: Get report by ID (Moderator/Admin)
      operationId: getReportByID
      parameters:
        - description: Report ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/moderationReportResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/reports/{id}/actions:
    post:
      description: Process a report with a moderation action (Moderator/Admin)
      operationId: processReport
      parameters:
        - description: Report ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/ModerationActionRequest"
      responses:
        "200":
          $ref: "#/responses/moderationActionResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/stats:
    get:
      description: Get moderation statistics (Moderator/Admin)
      operationId: getModerationStats
      parameters:
        - description: Optional moderator ID
          in: query
          name: moderator_id
          type: string
          x-go-name: ModeratorID
      responses:
        "200":
          $ref: "#/responses/moderationStatsResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/threads/{id}:
    post:
      description: Moderate a thread (Moderator/Admin)
      operationId: moderateThread
      parameters:
        - description: Thread ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/users/{id}/ban:
    post:
      description: Ban a user (Moderator/Admin)
      operationId: banUser
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/UserBanRequest"
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/users/{id}/moderator-notes:
    get:
      description: List moderator notes for a user (Moderator/Admin)
      operationId: listModeratorNotes
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - description: Page number
          format: int64
          in: query
          name: page
          type: integer
          x-go-name: Page
        - description: Page size (limit)
          format: int64
          in: query
          name: limit
          type: integer
          x-go-name: Limit
      responses:
        "200":
          $ref: "#/responses/moderatorNotesListResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
    post:
      description: Create a moderator note for a user (Moderator/Admin)
      operationId: createModeratorNote
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/CreateModeratorNoteRequest"
      responses:
        "201":
          $ref: "#/responses/successResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /moderation/users/{id}/unban:
    post:
      description: Unban a user (Moderator/Admin)
      operationId: unbanUser
      parameters:
        - description: User ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
  /replies/{id}:
    delete:
      description: Delete reply (Author or Moderator/Admin)
      operationId: deleteReply
      parameters:
        - description: Reply ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - replies
    get:
      description: Get reply by ID
      operationId: getReplyByID
      parameters:
        - description: Reply ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
      responses:
        "200":
          $ref: "#/responses/replyResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - replies
    patch:
      description: Update reply (Author or Moderator/Admin)
      operationId: updateReply
      parameters:
        - description: Reply ID
          in: path
          name: id
          required: true
          type: string
          x-go-name: ID
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/UpdateReplyRequest"
      responses:
        "200":
          $ref: "#/responses/replyResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "403":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - replies
  /reports:
    post:
      description: Submit a content report (Authenticated)
      operationId: createReport
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/CreateReportRequest"
      responses:
        "201":
          $ref: "#/responses/reportResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - moderation
...
  /threads:
    get:
      description: List threads
      operationId: listThreads
      parameters:
        - description: Page number (default 1)
          format: int64
          in: query
          name: page
          type: integer
          x-go-name: Page
        - description: Page size (default 20, max 100)
          format: int64
          in: query
          name: pageSize
          type: integer
          x-go-name: PageSize
      responses:
        "200":
          $ref: "#/responses/threadListResponse"
        "500":
          $ref: "#/responses/errorResponse"
      tags:
        - threads
    post:
      description: Create thread (Authenticated)
      operationId: createThread
      parameters:
        - in: body
          name: Body
          required: true
          schema:
            $ref: "#/definitions/CreateThreadRequest"
      responses:
        "201":
          $ref: "#/responses/threadResponse"
        "400":
          $ref: "#/responses/errorResponse"
        "401":
          $ref: "#/responses/errorResponse"
        "404":
          $ref: "#/responses/errorResponse"
        "409":
          $ref: "#/responses/errorResponse"
        "422":
          $ref: "#/responses/errorResponse"
        "500":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - threads
...
  /users/me/role:
    get:
      description: Get my role info (Authenticated)
      operationId: getRoleInfo
      responses:
        "200":
          $ref: "#/responses/successResponse"
        "401":
          $ref: "#/responses/errorResponse"
      security:
        - Bearer:
            - "[]"
      tags:
        - users
responses:
...
  errorResponse:
    description: Generic error envelope
    schema:
      $ref: "#/definitions/ErrorResponse"
...
  successResponse:
    description: Generic success envelope
    schema:
      $ref: "#/definitions/SuccessResponse"
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: Authorization
    description: "JWT token in the format 'Bearer {token}'"
swagger: "2.0"